+++ NEWS.md
@@ -1,3 +1,10 @@
+## ShiftedArrays 0.4 release notes
+
+### Breaking changes
+
+- Now conversion of `AbstractArray{<:ShiftedArray}` to `Array` or `OffsetArray` is done via `ShiftedArray.to_array` and `ShiftedArray.to_offsetarray` respectively rather than `Base.convert` to avoid type piracy.
+
+
 ## ShiftedArrays 0.3.1 release notes
 
 ### New features